+++ test/Spec.hs
@@ -1,4 +1,6 @@
 import Year2018.Day1Spec
+import Year2018.Day2Spec
 
 main :: IO ()
 main = do day1specs
+          day2specs