+++ ChangeLog.md
@@ -2,7 +2,11 @@
 
 ## Unreleased changes
 
-### 01.12.2018
+Nothing is actually released.
+
+## 01.12.2018
 
 * Add Code for Day 1
 * Add Specs for Day 1
+* Add String Interpolation
+* Add (empty) Specs for Day 2